remove old LOF score resampler
