+++ deepdrivewe/examples/synd_ntl9_lof/inference.py
@@ -20,9 +20,7 @@ from deepdrivewe.ai import ConvolutionalVAE
 from deepdrivewe.ai import ConvolutionalVAEConfig
 from deepdrivewe.binners import RectilinearBinner
 from deepdrivewe.recyclers import LowRecycler
-
-# from deepdrivewe.resamplers import LOFLowResampler
-from deepdrivewe.resamplers.lof_v2 import LOFLowResamplerV2
+from deepdrivewe.resamplers import LOFLowResampler
 
 
 class InferenceConfig(BaseModel):
@@ -285,7 +283,7 @@ def run_inference(
     )
 
     # Resamlpe the ensemble
-    resampler = LOFLowResamplerV2(
+    resampler = LOFLowResampler(
         consider_for_resampling=config.consider_for_resampling,
         max_resamples=config.max_resamples,
         max_allowed_weight=config.max_allowed_weight,
